+++ VaporShell/Public/SDK Wrappers/CloudFormation/Update-VSStackSet.ps1
@@ -103,6 +103,7 @@ function Update-VSStackSet {
                 }
                 Tags {
                     $tagList = New-Object 'System.Collections.Generic.List[Amazon.CloudFormation.Model.Tag]'
+                    $tagList.Add((VSStackTag -Key BuiltWith -Value VaporShell))
                     if ($null -ne $Tags) {
                         foreach ($key in $Tags.Keys) {
                             $tagList.Add((VSStackTag -Key $key -Value $Tags[$key]))