+++ psake.ps1
@@ -33,22 +33,7 @@ task Init {
     Set-Location $ProjectRoot
     "Build System Details:"
     Get-Item ENV:BH*
-    if ($env:APPVEYOR) {
-        Get-Item ENV:APPVEYOR*
-    }
-    "`n"
-
-    "    Installing the latest version of Pester"
-    Install-Module -Name Pester -Repository PSGallery -Scope CurrentUser -AllowClobber -SkipPublisherCheck -Confirm:$false -ErrorAction Stop -Force -Verbose:$false
-    Import-Module -Name Pester -Verbose:$false -Force
-    "    Installing the latest version of PSScriptAnalyzer"
-    Install-Module -Name PSScriptAnalyzer -Repository PSGallery -Scope CurrentUser -AllowClobber -SkipPublisherCheck -Confirm:$false -ErrorAction Stop -Force -Verbose:$false
-    Import-Module -Name PSScriptAnalyzer -Verbose:$false -Force
-    if ($env:BHBuildSystem -eq 'VSTS' -and $PSVersionTable.PSVersion.Major -lt 6) {
-        "    Installing the latest version of Coveralls"
-        Install-Module Coveralls -Repository PSGallery -Scope CurrentUser -ErrorAction Stop -Force -Confirm:$false -Verbose:$false
-        Import-Module Coveralls -Force -Verbose:$false
-    }
+    Get-Item ENV:BUILD*
     if ($env:BHProjectName -cne $moduleName) {
         $env:BHProjectName = $moduleName
     }
@@ -198,7 +183,15 @@ Export-ModuleMember -Function (Get-Command -Module VaporShell.DSL).Name -Variabl
     Get-ChildItem $outputModVerDir | Format-Table -Autosize
 } -description 'Compiles module from source'
 
+Task Import -Depends Compile {
+    '    Testing import of compiled module'
+    Import-Module (Join-Path $outputModVerDir "$($env:BHProjectName).psd1")
+} -description 'Imports the newly compiled module'
+
 $pesterScriptBlock = {
+    "    Installing the latest version of Pester"
+    Install-Module -Name Pester -Repository PSGallery -Scope CurrentUser -AllowClobber -SkipPublisherCheck -Confirm:$false -ErrorAction Stop -Force -Verbose:$false
+    Import-Module -Name Pester -Verbose:$false -Force
     "    Getting correctly cased FullName for $outputModDir..."
     $outputModDir = (Get-ChildItem (Split-Path $outputModDir -Parent) -Directory | Where-Object {$_.Name -eq (Get-Item $outputModDir).Name}).FullName
     "    FullName resolved to $outputModDir..."
@@ -224,29 +217,21 @@ $pesterScriptBlock = {
     Remove-Module $ENV:BHProjectName -ErrorAction SilentlyContinue -Verbose:$false
     Import-Module -Name $outputModDir -Force -Verbose:$false
     $testResultsXml = Join-Path -Path $outputDir -ChildPath $TestFile
-    $coverage = @{}
-    if ($PSVersionTable.PSVersion.Major -lt 6) {
-        $coveredFunctions = @()
-        "$($env:BHPSModulePath)\Public\Condition Functions\*","$($env:BHPSModulePath)\Public\Intrinsic Functions\*","$($env:BHPSModulePath)\Public\Primary Functions\*","$($env:BHPSModulePath)\Public\Transform\*","$($env:BHPSModulePath)\Public\*-Vaporshell.ps1" | Foreach-Object {
-            foreach ($item in (Get-Item $_)) {
-                $coveredFunctions += @{
-                    Path = (Join-Path $outputModVerDir "$($env:BHProjectName).psm1")
-                    Function = $item.BaseName
-                }
-            }
-        }
-        # Commented out due to extra time this takes when running against the compiled module
-        # $coverage['CodeCoverage'] = $coveredFunctions
+    $pesterParams = @{
+        OutputFormat = 'NUnitXml'
+        OutputFile   = $testResultsXml
+        PassThru     = $true
+        Path         = $tests
     }
-    '    Invoking Pester...'
-    $testResults = Invoke-Pester -Path $tests -PassThru -OutputFile $testResultsXml -OutputFormat NUnitXml #@coverage
+    if ($global:ExcludeTag) {
+        $pesterParams['ExcludeTag'] = $global:ExcludeTag
+        "    Invoking Pester and excluding tag(s) [$($global:ExcludeTag -join ', ')]..."
+    }
+    else {
+        '    Invoking Pester...'
+    }
+    $testResults = Invoke-Pester @pesterParams
     '    Pester invocation complete!'
-    if ($PSVersionTable.PSVersion.Major -lt 6 -and $null -ne $env:Coveralls -and $coverage.Keys -contains 'CodeCoverage') {
-        '    Uploading Code Coverage to Coveralls...'
-        $coverage = Format-Coverage -PesterResults $TestResults -CoverallsApiToken $env:Coveralls -BranchName $ENV:APPVEYOR_REPO_BRANCH -Verbose
-        Publish-Coverage -Coverage $coverage -Verbose
-    }
-
     if ($testResults.FailedCount -gt 0) {
         $testResults | Format-List
         Write-Error -Message 'One or more Pester tests failed. Build cannot continue!'
@@ -255,9 +240,9 @@ $pesterScriptBlock = {
     $env:PSModulePath = $origModulePath
 }
 
-task Pester -Depends Compile $pesterScriptBlock -description 'Run Pester tests'
+task Pester -Depends Import $pesterScriptBlock -description 'Run Pester tests'
 
-task PesterOnly -Depends Init $pesterScriptBlock -description 'Run Pester tests only (no Clean/Compile)'
+task PesterOnly -Depends Update $pesterScriptBlock -description 'Run Pester tests only (no Clean/Compile)'
 
 task Analyze -Depends Pester {
     $analysis = Invoke-ScriptAnalyzer -Path "$PSScriptRoot\$($env:BHProjectName)" -Recurse -Verbose:$false
@@ -279,8 +264,88 @@ task Analyze -Depends Pester {
     }
 } -description 'Run PSScriptAnalyzer'
 
-Task Deploy -Depends Compile {
-    if ($env:BHBuildSystem -eq 'VSTS' -and $env:BHCommitMessage -match '!deploy' -and $env:BHBranchName -eq "master") {
+Task Deploy -Depends Init {
+    function Publish-GitHubRelease {
+        <#
+            .SYNOPSIS
+            Publishes a release to GitHub Releases. Borrowed from https://www.herebedragons.io/powershell-create-github-release-with-artifact
+        #>
+        [CmdletBinding()]
+        Param (
+            [parameter(Mandatory = $true)]
+            [String]
+            $VersionNumber,
+            [parameter(Mandatory = $false)]
+            [String]
+            $CommitId = 'master',
+            [parameter(Mandatory = $true)]
+            [String]
+            $ReleaseNotes,
+            [parameter(Mandatory = $true)]
+            [ValidateScript( {Test-Path $_})]
+            [String]
+            $ArtifactPath,
+            [parameter(Mandatory = $true)]
+            [String]
+            $GitHubUsername,
+            [parameter(Mandatory = $true)]
+            [String]
+            $GitHubRepository,
+            [parameter(Mandatory = $true)]
+            [String]
+            $GitHubApiKey,
+            [parameter(Mandatory = $false)]
+            [Switch]
+            $PreRelease,
+            [parameter(Mandatory = $false)]
+            [Switch]
+            $Draft
+        )
+        $releaseData = @{
+            tag_name         = [string]::Format("v{0}", $VersionNumber)
+            target_commitish = $CommitId
+            name             = [string]::Format("$($env:BHProjectName) v{0}", $VersionNumber)
+            body             = $ReleaseNotes
+            draft            = [bool]$Draft
+            prerelease       = [bool]$PreRelease
+        }
+
+        $auth = 'Basic ' + [Convert]::ToBase64String([Text.Encoding]::ASCII.GetBytes($gitHubApiKey + ":x-oauth-basic"))
+
+        $releaseParams = @{
+            Uri         = "https://api.github.com/repos/$GitHubUsername/$GitHubRepository/releases"
+            Method      = 'POST'
+            Headers     = @{
+                Authorization = $auth
+            }
+            ContentType = 'application/json'
+            Body        = (ConvertTo-Json $releaseData -Compress)
+        }
+        [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+        $result = Invoke-RestMethod @releaseParams
+        $uploadUri = $result | Select-Object -ExpandProperty upload_url
+        $uploadUri = $uploadUri -creplace '\{\?name,label\}'
+        $artifact = Get-Item $ArtifactPath
+        $uploadUri = $uploadUri + "?name=$($artifact.Name)"
+        $uploadFile = $artifact.FullName
+
+        $uploadParams = @{
+            Uri         = $uploadUri
+            Method      = 'POST'
+            Headers     = @{
+                Authorization = $auth
+            }
+            ContentType = 'application/zip'
+            InFile      = $uploadFile
+        }
+        $result = Invoke-RestMethod @uploadParams
+    }
+    if (($env:BHBuildSystem -eq 'VSTS' -and $env:BHCommitMessage -match '!deploy' -and $env:BHBranchName -eq "master") -or $global:ForceDeploy -eq $true) {
+        if ($null -eq (Get-Module PoshTwit -ListAvailable)) {
+            "    Installing PoshTwit module..."
+            Install-Module PoshTwit -Scope CurrentUser
+        }
+        Import-Module PoshTwit -Verbose:$false
         # Load the module, read the exported functions, update the psd1 FunctionsToExport
         $commParsed = $env:BHCommitMessage | Select-String -Pattern '\sv\d\.\d\.\d\s'
         if ($commParsed) {
@@ -322,9 +387,83 @@ Task Deploy -Depends Compile {
         }
         # Bump the module version
         if ($versionToDeploy) {
-            Update-Metadata -Path (Join-Path $outputModVerDir "$($env:BHProjectName).psd1") -PropertyName ModuleVersion -Value $versionToDeploy
-            "    Publishing version [$($versionToDeploy)] to PSGallery..."
-            Publish-Module -Path $outputModVerDir -NuGetApiKey $env:NugetApiKey -Repository PSGallery
+            try {
+                if ($ENV:BHBuildSystem -eq 'VSTS' -and -not [String]::IsNullOrEmpty($env:NugetApiKey)) {
+                    "    Publishing version [$($versionToDeploy)] to PSGallery..."
+                    Update-Metadata -Path (Join-Path $outputModVerDir "$($env:BHProjectName).psd1") -PropertyName ModuleVersion -Value $versionToDeploy
+                    Publish-Module -Path $outputModVerDir -NuGetApiKey $env:NugetApiKey -Repository PSGallery
+                    "    Deployment successful!"
+                }
+                else {
+                    "    [SKIPPED] Deployment of version [$($versionToDeploy)] to PSGallery"
+                }
+                $commitId = git rev-parse --verify HEAD
+                if (-not [String]::IsNullOrEmpty($env:GitHubPAT)) {
+                    "    Creating Release ZIP..."
+                    $zipPath = [System.IO.Path]::Combine($PSScriptRoot,"$($env:BHProjectName).zip")
+                    if (Test-Path $zipPath) {
+                        Remove-Item $zipPath -Force
+                    }
+                    Add-Type -Assembly System.IO.Compression.FileSystem
+                    [System.IO.Compression.ZipFile]::CreateFromDirectory($outputModDir,$zipPath)
+                    "    Publishing Release v$($versionToDeploy) @ commit Id [$($commitId)] to GitHub..."
+                    $ReleaseNotes = "# Changelog`n`n"
+                    $ReleaseNotes += (git log -1 --pretty=%B | Select-Object -Skip 2) -join "`n"
+                    $ReleaseNotes += "`n`n***`n`n# Instructions`n`n"
+                    $ReleaseNotes += @"
+1. [Click here](https://github.com/scrthq/$($env:BHProjectName)/releases/download/v$($versionToDeploy.ToString())/$($env:BHProjectName).zip) to download the *$($env:BHProjectName).zip* file attached to the release.
+2. **If on Windows**: Right-click the downloaded zip, select Properties, then unblock the file.
+    > _This is to prevent having to unblock each file individually after unzipping._
+3. Unzip the archive.
+4. (Optional) Place the module folder somewhere in your ``PSModulePath``.
+    > _You can view the paths listed by running the environment variable ```$env:PSModulePath``_
+5. Import the module, using the full path to the PSD1 file in place of ``$($env:BHProjectName)`` if the unzipped module folder is not in your ``PSModulePath``:
+    ``````powershell
+    # In `$env:PSModulePath
+    Import-Module $($env:BHProjectName)
+
+    # Otherwise, provide the path to the manifest:
+    Import-Module -Path C:\MyPSModules\$($env:BHProjectName)\$($versionToDeploy.ToString())\$($env:BHProjectName).psd1
+    ``````
+"@
+                    $gitHubParams = @{
+                        VersionNumber    = $versionToDeploy.ToString()
+                        CommitId         = $commitId
+                        ReleaseNotes     = $ReleaseNotes
+                        ArtifactPath     = $zipPath
+                        GitHubUsername   = 'scrthq'
+                        GitHubRepository = $env:BHProjectName
+                        GitHubApiKey     = $env:GitHubPAT
+                        Draft            = $false
+                    }
+                    Publish-GitHubRelease @gitHubParams
+                    "    Release creation successful!"
+                }
+                else {
+                    "    [SKIPPED] Publishing Release v$($versionToDeploy) @ commit Id [$($commitId)] to GitHub"
+                }
+                if ($ENV:BHBuildSystem -eq 'VSTS' -and -not [String]::IsNullOrEmpty($env:TwitterAccessSecret) -and -not [String]::IsNullOrEmpty($env:TwitterAccessToken) -and -not [String]::IsNullOrEmpty($env:TwitterConsumerKey) -and -not [String]::IsNullOrEmpty($env:TwitterConsumerSecret)) {
+                    "    Publishing tweet about new release..."
+                    $manifest = Import-PowerShellDataFile -Path (Join-Path $outputModVerDir "$($env:BHProjectName).psd1")
+                    $text = "$($env:BHProjectName) v$($versionToDeploy) is now available on the #PSGallery! https://www.powershellgallery.com/packages/$($env:BHProjectName) #PowerShell"
+                    $manifest.PrivateData.PSData.Tags | Foreach-Object {
+                        $text += " #$($_)"
+                    }
+                    if ($text.Length -gt 280) {
+                        "    Trimming [$($text.Length - 280)] extra characters from tweet text to get to 280 character limit..."
+                        $text = $text.Substring(0,280)
+                    }
+                    "    Tweet text: $text"
+                    Publish-Tweet -Tweet $text -ConsumerKey $env:TwitterConsumerKey -ConsumerSecret $env:TwitterConsumerSecret -AccessToken $env:TwitterAccessToken -AccessSecret $env:TwitterAccessSecret
+                    "    Tweet successful!"
+                }
+                else {
+                    "    [SKIPPED] Twitter update of new release"
+                }
+            }
+            catch {
+                Write-Error $_ -ErrorAction Stop
+            }
         }
         else {
             Write-Host -ForegroundColor Yellow "No module version matched! Negating deployment to prevent errors"
@@ -335,4 +474,6 @@ Task Deploy -Depends Compile {
     else {
         Write-Host -ForegroundColor Magenta "Build system is not VSTS, commit message does not contain '!deploy' and/or branch is not 'master' -- skipping module update!"
     }
-} -description 'Deploy module to PSGallery'
+} -description 'Deploy module to PSGallery' -preaction {
+    Import-Module -Name $outputModDir -Force -Verbose:$false
+}